Move status icon into separate directory
* Remove ButtonGroup export
