+++ awx/ui_next/src/components/CodeMirrorInput/index.js
@@ -4,4 +4,3 @@ export default CodeMirrorInput;
 export { default as VariablesDetail } from './VariablesDetail';
 export { default as VariablesInput } from './VariablesInput';
 export { default as VariablesField } from './VariablesField';
-export { default as ButtonGroup } from './ButtonGroup';
