+++ awx/ui_next/src/components/Sparkline/index.js
@@ -1,2 +1,1 @@
-export { default as Sparkline } from './Sparkline';
-export { default as StatusIcon } from './StatusIcon';
+export { default } from './Sparkline';
