+++ awx/ui_next/src/components/StatusIcon/StatusIcon.test.jsx
