+++ awx/ui_next/src/components/StatusIcon/index.js
@@ -1,0 +1,1 @@
+export { default } from './StatusIcon';
