+++ awx/ui_next/src/components/Workflow/WorkflowActionTooltipItem.test.jsx
@@ -4,7 +4,7 @@ import WorkflowActionTooltipItem from './WorkflowActionTooltipItem';
 
 describe('WorkflowActionTooltipItem', () => {
   test('successfully mounts', () => {
-    const wrapper = mount(<WorkflowActionTooltipItem />);
+    const wrapper = mount(<WorkflowActionTooltipItem id="node" />);
     expect(wrapper).toHaveLength(1);
   });
 });
