+++ awx/ui_next/src/screens/Host/HostList/HostListItem.jsx
@@ -16,7 +16,7 @@ import {
 import { Link } from 'react-router-dom';
 import { PencilAltIcon } from '@patternfly/react-icons';
 
-import { Sparkline } from '@components/Sparkline';
+import Sparkline from '@components/Sparkline';
 import { Host } from '@types';
 import styled from 'styled-components';
